backend/requirements.txt
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask==1.0.3
Jinja2==2.10.1
Werkzeug==0.15.4
flasgger==0.9.2
itsdangerous==1.1.0
pytz==2019.1
requests==2.22.0
setproctitle==1.1.10
six==1.12.0
dxfgrabber==1.0.1
numpy==1.18.1
Shapely==1.6.4.post2
flask-cors==3.0.8
py-kim==1.2.4